fix: missing header dropdown on mobile
